+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.2.3
+* Optimize permission request logic on android
+## 1.2.2+2
+* fix crash when requestPermission
+## 1.2.2+1
+* migrate to androidx
+## 1.2.2
+* ask for permission when needed
+## 1.2.1
+* fix somethings
 ## 1.2.0
 * Add custom parameters,"type" android,"uti" iOS
 ## 1.1.1